requirements.txt
langchain_anthropic==0.1.19
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.1
streamlit==1.36.0
